requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.7.0rc1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.8
django-allauth==0.42.0
django-crispy-forms==1.9.1
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
urllib3==1.25.9
